+++ app/src/main/java/com/techcourse/Application.java
@@ -1,10 +1,10 @@
 package com.techcourse;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
+import com.interface21.webmvc.servlet.TomcatStarter;
 import java.io.IOException;
 import java.util.stream.Stream;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class Application {
