+++ mvc/src/main/java/com/interface21/webmvc/servlet/TomcatStarter.java
@@ -1,13 +1,12 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet;
 
+import java.io.File;
 import org.apache.catalina.Context;
 import org.apache.catalina.LifecycleException;
 import org.apache.catalina.connector.Connector;
 import org.apache.catalina.core.StandardContext;
 import org.apache.catalina.startup.Tomcat;
 import org.apache.tomcat.util.scan.StandardJarScanner;
-
-import java.io.File;
 
 public class TomcatStarter {
